app/requirements-dev.txt
biopython
Django
django-cors-headers
bs4
django-filter
djangorestframework
drf-yasg
mysqlclient
python-dotenv
django-debug-toolbar
django-debug-toolbar-force
pandas
rcsbsearchapi
whitenoise